perplexity_deep_research/pyproject.toml
[project]
name = "perplexity-deep-research"
version = "0.1.0"
description = "MCP server for Perplexity Sonar Deep Research"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "mcp[cli]>=1.18.0",
    "perplexityai>=0.20.0",
    "python-dotenv>=1.0.0",
]
